requirements.txt
prettytable
aiohttp
PyInquirer
Colr
python-cfonts